# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Full studio settings blob — all sliders, toggles, and colors
 */
export interface FidgetSettingsBlob {
  [key: string]: unknown;
}

export interface Project {
  id: number;
  userId: string;
  name: string;
  /** Raw SVG content */
  svgData: string;
  /** Extrusion depth in mm */
  extrudeDepth: number;
  /** Keycap square hole size in mm */
  keycapSize: number;
  /** Circular peg radius in mm */
  pegRadius: number;
  /** Full studio settings blob */
  settings?: FidgetSettingsBlob | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  svgData: string;
  extrudeDepth?: number;
  keycapSize?: number;
  pegRadius?: number;
  /** Full studio settings blob */
  settings?: FidgetSettingsBlob;
}

export interface UpdateProjectBody {
  name?: string;
  svgData?: string;
  extrudeDepth?: number;
  keycapSize?: number;
  pegRadius?: number;
  /** Full studio settings blob */
  settings?: FidgetSettingsBlob;
}

export interface ProjectStats {
  totalProjects: number;
  totalExports: number;
  mostRecentProject?: Project | null;
}