# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SVG Fidget Toy Creator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Fidget toy project operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user's projects
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new fidget toy project
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized

  /projects/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project statistics for the user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Project stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"
        "401":
          description: Unauthorized

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    FidgetSettingsBlob:
      type: object
      description: Full studio settings blob — all sliders, toggles, and colors
      additionalProperties: true

    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        svgData:
          type: string
          description: Raw SVG content
        extrudeDepth:
          type: number
          description: Extrusion depth in mm
        keycapSize:
          type: number
          description: Keycap square hole size in mm
        pegRadius:
          type: number
          description: Circular peg radius in mm
        settings:
          $ref: "#/components/schemas/FidgetSettingsBlob"
          nullable: true
          description: Full studio settings blob
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - svgData
        - extrudeDepth
        - keycapSize
        - pegRadius
        - createdAt
        - updatedAt

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        svgData:
          type: string
        extrudeDepth:
          type: number
          default: 4
        keycapSize:
          type: number
          default: 14
        pegRadius:
          type: number
          default: 3.5
        settings:
          $ref: "#/components/schemas/FidgetSettingsBlob"
          description: Full studio settings blob
      required:
        - name
        - svgData

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        svgData:
          type: string
        extrudeDepth:
          type: number
        keycapSize:
          type: number
        pegRadius:
          type: number
        settings:
          $ref: "#/components/schemas/FidgetSettingsBlob"
          description: Full studio settings blob

    ProjectStats:
      type: object
      properties:
        totalProjects:
          type: integer
        totalExports:
          type: integer
        mostRecentProject:
          $ref: "#/components/schemas/Project"
          nullable: true
      required:
        - totalProjects
        - totalExports