# lib/api-zod/src/generated/types/fidgetSettingsBlob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Full studio settings blob — all sliders, toggles, and colors
 */
export interface FidgetSettingsBlob {
  [key: string]: unknown;
}