# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { FidgetSettingsBlob } from "./fidgetSettingsBlob";

export interface Project {
  id: number;
  userId: string;
  name: string;
  /** Raw SVG content */
  svgData: string;
  /** Extrusion depth in mm */
  extrudeDepth: number;
  /** Keycap square hole size in mm */
  keycapSize: number;
  /** Circular peg radius in mm */
  pegRadius: number;
  /** Full studio settings blob */
  settings?: FidgetSettingsBlob | null;
  createdAt: Date;
  updatedAt: Date;
}