# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { FidgetSettingsBlob } from "./fidgetSettingsBlob";

export interface UpdateProjectBody {
  name?: string;
  svgData?: string;
  extrudeDepth?: number;
  keycapSize?: number;
  pegRadius?: number;
  /** Full studio settings blob */
  settings?: FidgetSettingsBlob;
}