# lib/db/src/schema/projects.ts
import { pgTable, serial, text, real, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const projectsTable = pgTable("projects", {
  id: serial("id").primaryKey(),
  userId: text("user_id").notNull(),
  name: text("name").notNull(),
  svgData: text("svg_data").notNull(),
  extrudeDepth: real("extrude_depth").notNull().default(4),
  keycapSize: real("keycap_size").notNull().default(14),
  pegRadius: real("peg_radius").notNull().default(3.5),
  settings: jsonb("settings"),
  exportCount: serial("export_count").notNull(),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertProjectSchema = createInsertSchema(projectsTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
  exportCount: true,
});

export type InsertProject = z.infer<typeof insertProjectSchema>;
export type Project = typeof projectsTable.$inferSelect;